Update Models and Assignd Controller - 20171223

+++ priland-api/Controllers/MstHouseModelController.cs
@@ -11,12 +11,13 @@ using Microsoft.AspNet.Identity;
 namespace priland_api.Controllers
 {
     [Authorize]
+    [RoutePrefix("api/MstHouseModel")]
     public class MstHouseModelController : ApiController
     {
         private Data.FilbrokerDBDataContext db = new Data.FilbrokerDBDataContext();
          
         //List
-        [HttpGet, Route("api/MstHouseModel/List")]
+        [HttpGet, Route("List")]
         public List<Models.MstHouseModel> GetMstHouseModel()
         {
             var MstHouseModel = from d in db.MstHouseModels
@@ -27,8 +28,10 @@ namespace priland_api.Controllers
                                     HouseModelCode = d.HouseModelCode,
                                     HouseModel = d.HouseModel,
                                     ProjectId = d.ProjectId,
+                                    Project=d.MstProject.Project,
                                     TFA = d.TFA,
                                     Price = d.Price,
+                                    IsLocked=d.IsLocked,
                                     CreatedBy = d.CreatedBy,
                                     CreatedDateTime = d.CreatedDateTime.ToShortDateString(),
                                     UpdatedBy = d.UpdatedBy,
@@ -36,8 +39,9 @@ namespace priland_api.Controllers
                                 };
             return MstHouseModel.ToList();
         }
+
         //Get Record Detail
-        [HttpGet, Route("api/MstHouseModel/Detail/{id}")]
+        [HttpGet, Route("Detail/{id}")]
         public Models.MstHouseModel GetMstHouseModelId(string id)
         {
             var MstHouseModelData = from d in db.MstHouseModels
@@ -50,6 +54,7 @@ namespace priland_api.Controllers
                                         ProjectId = d.ProjectId,
                                         TFA = d.TFA,
                                         Price = d.Price,
+                                        IsLocked=d.IsLocked,
                                         CreatedBy = d.CreatedBy,
                                         CreatedDateTime = d.CreatedDateTime.ToShortDateString(),
                                         UpdatedBy = d.UpdatedBy,
@@ -59,27 +64,38 @@ namespace priland_api.Controllers
         }
 
         //Add New Record
-        [HttpPost, Route("api/MstHouseModel/Add")]
+        [HttpPost, Route("Add")]
         public int PostMstHouseModel(Models.MstHouseModel addMstHouseModel)
         {
             try
             {
-                Data.MstHouseModel newMstHouseModel = new Data.MstHouseModel();
+                var currentUser = from d in db.MstUsers
+                                  where d.AspNetId == User.Identity.GetUserId()
+                                  select d;
+                if (currentUser.Any())
+                {
+                    Data.MstHouseModel newMstHouseModel = new Data.MstHouseModel()
+                    {
+                        HouseModelCode = addMstHouseModel.HouseModelCode,
+                        HouseModel = addMstHouseModel.HouseModel,
+                        ProjectId = addMstHouseModel.ProjectId,
+                        TFA = addMstHouseModel.TFA,
+                        Price = addMstHouseModel.Price,
+                        IsLocked=false,
+                        CreatedBy = currentUser.FirstOrDefault().Id,
+                        CreatedDateTime = DateTime.Now,
+                        UpdatedBy =currentUser.FirstOrDefault().Id,
+                        UpdatedDateTime = DateTime.Now
+                    };
+                    db.MstHouseModels.InsertOnSubmit(newMstHouseModel);
+                    db.SubmitChanges();
 
-                newMstHouseModel.HouseModelCode = addMstHouseModel.HouseModelCode;
-                newMstHouseModel.HouseModel = addMstHouseModel.HouseModel;
-                newMstHouseModel.ProjectId = addMstHouseModel.ProjectId;
-                newMstHouseModel.TFA = addMstHouseModel.TFA;
-                newMstHouseModel.Price = addMstHouseModel.Price;
-                newMstHouseModel.CreatedBy = addMstHouseModel.CreatedBy;
-                newMstHouseModel.CreatedDateTime = Convert.ToDateTime(addMstHouseModel.CreatedDateTime);
-                newMstHouseModel.UpdatedBy = addMstHouseModel.UpdatedBy;
-                newMstHouseModel.UpdatedDateTime = Convert.ToDateTime(addMstHouseModel.UpdatedDateTime);
-
-                db.MstHouseModels.InsertOnSubmit(newMstHouseModel);
-                db.SubmitChanges();
-
-                return newMstHouseModel.Id;
+                    return newMstHouseModel.Id;
+                }
+                else
+                {
+                    return 0;
+                }
             }
             catch (Exception e)
             {
@@ -87,10 +103,9 @@ namespace priland_api.Controllers
                 return 0;
             }
         }
-
-
+        
         //Delete Record
-        [HttpDelete, Route("api/MstHouseModel/Delete/{id}")]
+        [HttpDelete, Route("Delete/{id}")]
         public HttpResponseMessage DeleteMstHouseModel(string id)
         {
             try
@@ -98,10 +113,17 @@ namespace priland_api.Controllers
                 var MstHouseModelData = from d in db.MstHouseModels where d.Id == Convert.ToInt32(id) select d;
                 if (MstHouseModelData.Any())
                 {
-                    db.MstHouseModels.DeleteOnSubmit(MstHouseModelData.First());
-                    db.SubmitChanges();
+                    if (MstHouseModelData.FirstOrDefault().IsLocked == false)
+                    {
+                        db.MstHouseModels.DeleteOnSubmit(MstHouseModelData.First());
+                        db.SubmitChanges();
 
-                    return Request.CreateResponse(HttpStatusCode.OK);
+                        return Request.CreateResponse(HttpStatusCode.OK);
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest);
+                    }
                 }
                 else
                 {
@@ -115,30 +137,92 @@ namespace priland_api.Controllers
             }
         }
 
-        //Update Record Data
-        [HttpPut, Route("api/MstHouseModel/Update/{id}")]
-        public HttpResponseMessage UpdateHouseModel(string id, Models.MstHouseModel UpdateMstHouseModel)
+        //Save
+        [HttpPut, Route("Save")]
+        public HttpResponseMessage SaveMstHouseModel(MstHouseModel housemodel)
         {
             try
             {
-                var MstHouseModelData = from d in db.MstHouseModels where d.Id == Convert.ToInt32(id) select d;
+                var MstHouseModelData = from d in db.MstHouseModels where d.Id == Convert.ToInt32(housemodel.Id) select d;
                 if (MstHouseModelData.Any())
                 {
-                    var UpdateHouseModelData = MstHouseModelData.FirstOrDefault();
+                    if (MstHouseModelData.First().IsLocked == false)
+                    {
+                        var currentUser = from d in db.MstUsers
+                                          where d.AspNetId == User.Identity.GetUserId()
+                                          select d;
 
-                    UpdateHouseModelData.HouseModelCode = UpdateMstHouseModel.HouseModelCode;
-                    UpdateHouseModelData.HouseModel = UpdateMstHouseModel.HouseModel;
-                    UpdateHouseModelData.ProjectId = UpdateMstHouseModel.ProjectId;
-                    UpdateHouseModelData.TFA = UpdateMstHouseModel.TFA;
-                    UpdateHouseModelData.Price = UpdateMstHouseModel.Price;
-                    UpdateHouseModelData.CreatedBy = UpdateMstHouseModel.CreatedBy;
-                    UpdateHouseModelData.CreatedDateTime = Convert.ToDateTime(UpdateMstHouseModel.CreatedDateTime);
-                    UpdateHouseModelData.UpdatedBy = UpdateMstHouseModel.UpdatedBy;
-                    UpdateHouseModelData.UpdatedDateTime = Convert.ToDateTime(UpdateMstHouseModel.UpdatedDateTime);
+                        if (currentUser.Any())
+                        {
+                            var UpdateMstHouseModelData = MstHouseModelData.FirstOrDefault();
 
-                    db.SubmitChanges();
+                            UpdateMstHouseModelData.HouseModelCode = housemodel.HouseModelCode;
+                            UpdateMstHouseModelData.HouseModel = housemodel.HouseModel;
+                            UpdateMstHouseModelData.ProjectId = housemodel.ProjectId;
+                            UpdateMstHouseModelData.TFA = housemodel.TFA;
+                            UpdateMstHouseModelData.Price = housemodel.Price;
+                            UpdateMstHouseModelData.UpdatedBy = currentUser.FirstOrDefault().Id;
+                            UpdateMstHouseModelData.UpdatedDateTime = DateTime.Now;
 
+                            db.SubmitChanges();
+
+                            return Request.CreateResponse(HttpStatusCode.OK);
+                        }
+                        else
+                        {
+                            return Request.CreateResponse(HttpStatusCode.BadRequest);
+                        }
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest);
+                    }
+                }
+                else
+                {
                     return Request.CreateResponse(HttpStatusCode.OK);
+                }
+            }
+            catch (Exception e)
+            {
+                Debug.WriteLine(e);
+                return Request.CreateResponse(HttpStatusCode.BadRequest);
+            }
+        }
+
+        //Lock Record Data
+        [HttpPut, Route("Lock")]
+        public HttpResponseMessage UpdateHouseModel(MstHouseModel UpdateMstHouseModel)
+        {
+            try
+            {
+                var MstHouseModelData = from d in db.MstHouseModels where d.Id == Convert.ToInt32(UpdateMstHouseModel.Id) select d;
+                if (MstHouseModelData.Any())
+                {
+                    var currentUser = from d in db.MstUsers
+                                      where d.AspNetId == User.Identity.GetUserId()
+                                      select d;
+                    if (currentUser.Any())
+                    {
+                        var UpdateHouseModelData = MstHouseModelData.FirstOrDefault();
+
+                        UpdateHouseModelData.HouseModelCode = UpdateMstHouseModel.HouseModelCode;
+                        UpdateHouseModelData.HouseModel = UpdateMstHouseModel.HouseModel;
+                        UpdateHouseModelData.ProjectId = UpdateMstHouseModel.ProjectId;
+                        UpdateHouseModelData.TFA = UpdateMstHouseModel.TFA;
+                        UpdateHouseModelData.Price = UpdateMstHouseModel.Price;
+                        UpdateHouseModelData.IsLocked = true;
+                        UpdateHouseModelData.UpdatedBy = currentUser.FirstOrDefault().Id;
+                        UpdateHouseModelData.UpdatedDateTime = DateTime.Now;
+
+                        db.SubmitChanges();
+
+                        return Request.CreateResponse(HttpStatusCode.OK);
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest);
+                    }
                 }
                 else
                 {
@@ -152,5 +236,48 @@ namespace priland_api.Controllers
             }
 
         }
+
+        //Lock Record Data
+        [HttpPut, Route("UnLock")]
+        public HttpResponseMessage UnLockHouseModel(MstHouseModel UnLockMstHouseModel)
+        {
+            try
+            {
+                var MstHouseModelData = from d in db.MstHouseModels where d.Id == Convert.ToInt32(UnLockMstHouseModel.Id) select d;
+                if (MstHouseModelData.Any())
+                {
+                    var currentUser = from d in db.MstUsers
+                                      where d.AspNetId == User.Identity.GetUserId()
+                                      select d;
+                    if (currentUser.Any())
+                    {
+                        var UpdateHouseModelData = MstHouseModelData.FirstOrDefault();
+
+                        UpdateHouseModelData.IsLocked = false;
+                        UpdateHouseModelData.UpdatedBy = currentUser.FirstOrDefault().Id;
+                        UpdateHouseModelData.UpdatedDateTime = DateTime.Now;
+
+                        db.SubmitChanges();
+
+                        return Request.CreateResponse(HttpStatusCode.OK);
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest);
+                    }
+                }
+                else
+                {
+                    return Request.CreateResponse(HttpStatusCode.OK);
+                }
+            }
+            catch (Exception e)
+            {
+                Debug.WriteLine(e);
+                return Request.CreateResponse(HttpStatusCode.BadRequest);
+            }
+
+        }
+
     }
 }
